Title: a의 b승, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240122/a의 b승/a-to-the-power-of-b.cpp
@@ -6,6 +6,6 @@ int main() {
     
     int a, b;
     cin >> a >> b;
-    cout << pow(a,b);
+    cout << (double)pow(a,b);
     return 0;
 }
